Category pages added
