+++ .byebug_history
@@ -1,4 +1,8 @@
 exit
+@pages
+exit
+@pages
+exit
 @jobs.count
 exit
 conitnue
